% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmpLoading1dPlot.R
\name{lmpLoading1dPlot}
\alias{lmpLoading1dPlot}
\title{Loadings represented on a line plot.}
\usage{
lmpLoading1dPlot(resLmpPcaEffects, effectNames = NULL, axes = c(1, 2), ...)
}
\arguments{
\item{resLmpPcaEffects}{A list corresponding to the output value of \code{\link{lmpPcaEffects}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{...}{Additional arguments to be passed to \code{\link{plotLine}}.}
}
\value{
A list of \code{ggplot} objects representing the loading plots.
}
\description{
Plots the loadings for each effect matrix from \code{\link{lmpPcaEffects}} outputs with line plots.
}
\details{
\code{lmpLoading1dPlot} is a wrapper of \code{\link{plotLine}}.
}
\examples{

data('UCH')
resLmpModelMatrix = lmpModelMatrix(UCH)
resLmpEffectMatrices = lmpEffectMatrices(resLmpModelMatrix)
resASCA = lmpPcaEffects(resLmpEffectMatrices)

lmpLoading1dPlot(resASCA, effectNames = c("Hippurate", "Citrate"))
}
