% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmpOutcomesReduct.R
\name{lmpOutcomesReduct}
\alias{lmpOutcomesReduct}
\title{Reduce the number of columns in the outcomes matrix}
\usage{
lmpOutcomesReduct(lmpDataList, nPC = NULL)
}
\arguments{
\item{lmpDataList}{A list containing the outcomes, the experimental design, the formula and the type of model (model = 'lm' for model linear and model = 'lmm' for model linear mixt).}

\item{nPC}{The number of first principal components}
}
\value{
A list with the 3 following named elements and the result of PCA :
\describe{
\item{\code{lmpDataList}}{a list with outcomes (scores of PCA), design, formula, type of model, old outcomes, loadings PCA and Boolean \emph{isReduct"} = TRUE }
\item{\code{resPCA}}{a list with result of PCA}
\item{\code{nPC}}{The number of first principal components}
}
}
\description{
Reduce the number of columns in the outcomes matrix with the PCA.
This function modify the \emph{lmpDataList} object by adding the loadings from PCA \emph{loadingsPCA} and a boolean \emph{isReduct} is TRUE to specify that the matrix has indeed been reduced.
\emph{outcomes} are replaced by the scores of the PCA, and a \emph{outcomesRaw} retains the initial outcomes.
It is possible to choose the number of principal components with the \emph{nPC} parameter. If this parameter is not specified, the function will take the first principal components to achieve a cumulative variance of 0.99.
}
