% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwBootstrapTests.R
\name{lmwBootstrapTests}
\alias{lmwBootstrapTests}
\title{Performs a test on the effects from the model}
\usage{
lmwBootstrapTests(
  resLmwEffectMatrices,
  nboot = 100,
  nCores = 2,
  verbose = FALSE
)
}
\arguments{
\item{resLmwEffectMatrices}{A list of 12 from \code{\link{lmwEffectMatrices}}.}

\item{nboot}{An integer with the number of iterations to perform.}

\item{nCores}{The number of cores to use for parallel execution.}

\item{verbose}{If \code{TRUE}, will display a message with the duration of execution.}
}
\value{
A list with the following elements:
\describe{
\item{\code{f.obs}}{A vector of size F with the F statistics calculated on the initial data for each model term.}
\item{\code{f.boot}}{A b × F matrix with the F statistics calculated for the bootstrap samples.}
\item{\code{p.values}}{A vector of size F with the p-value for each model effect.}
\item{\code{resultsTable}}{A 2 × F matrix with the p-value and the variance percentage for each model effect.}
}
}
\description{
Tests the significance of the effects from the model using bootstrap. This function is based on the outputs of \code{\link{lmwEffectMatrices}}. Tests on combined effects are also provided.
}
\examples{
 data('UCH')
 resLmwModelMatrix <- lmwModelMatrix(UCH)
 resLmwEffectMatrices <- lmwEffectMatrices(resLmwModelMatrix = resLmwModelMatrix)

 res <- lmwBootstrapTests(resLmwEffectMatrices = resLmwEffectMatrices, nboot=10, nCores=2, verbose = TRUE)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
