% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMWContributions.R
\name{LMWContributions}
\alias{LMWContributions}
\title{Summary of the contributions from each effect}
\usage{
LMWContributions(ResLMWtoPCAEffects, nPC = 5)
}
\arguments{
\item{ResLMWtoPCAEffects}{A ResLMWtoPCAEffects list from \code{\link{LMWtoPCAEffects}}}

\item{nPC}{The number of Principal Components to print in the tables}
}
\value{
A list of 3 :
\describe{
\item{\code{EffectTable}}{A matrix with the variance percentage of each PC for each effect}
\item{\code{ContribTable}}{A matrix with the contribution of each PC to the total variance}
\item{\code{Barplot}}{A barplot with the PCs which have the biggest contributions to the total variance}
}
}
\description{
Create a summary of the contribution of each effect and each of their Principal Component (PC) on the total variance. Additionally plots a graph with the ordered contributions.
}
\examples{
data('UCH')
ResLMWModelMatrix = LMWModelMatrix(UCH)
ResLMWEffectMatrices = LMWEffectMatrices(ResLMWModelMatrix)
ResLMWtoPCAEffects = LMWtoPCAEffects(ResLMWEffectMatrices, method="ASCA-E")
LMWContributions(ResLMWtoPCAEffects)
}
