% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwEffectPlot.R
\name{lmwEffectPlot}
\alias{lmwEffectPlot}
\title{Effect plot}
\usage{
lmwEffectPlot(
  resASCA,
  effectName,
  axes = 1,
  x,
  z = NULL,
  w = NULL,
  hline = 0,
  ...
)
}
\arguments{
\item{resASCA}{A list corresponding to the ASCA output value of \code{\link{lmwPcaEffects}}.}

\item{effectName}{Name of the effect matrix used for the scores.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{x}{A character string giving the \code{design} factor whose levels will form the x axis.}

\item{z}{A character string giving the \code{design} factor whose levels will form the traces.}

\item{w}{A character string giving the \code{design} factor whose levels will be used for the facet.}

\item{...}{}
}
\value{
An effect plot (ggplot).
}
\description{
Draws effects plot for the \code{\link{lmwPcaEffects}} function.
}
\details{
\code{lmwEffectPlot} is a wrapper of \code{\link{plotMeans}}.
}
\examples{

}
