% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwLoading1dPlot.R
\name{lmwLoading1dPlot}
\alias{lmwLoading1dPlot}
\title{Loadings represented on a line plot.}
\usage{
lmwLoading1dPlot(resLmwPcaEffects, effectNames = NULL, axes = c(1, 2), ...)
}
\arguments{
\item{resLmwPcaEffects}{A list corresponding to the output value of \code{\link{lmwPcaEffects}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{...}{Additional arguments to be passed to \code{\link{plotLine}}.}
}
\value{
A list of loading plots (ggplot).
}
\description{
Plots the loadings for each effect matrix from \code{\link{lmwPcaEffects}} outputs with line plots.
}
\details{
\code{lmwLoading1dPlot} is a wrapper of \code{\link{plotLine}}.
}
\examples{

data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resASCA = lmwPcaEffects(resLmwEffectMatrices)

lmwLoading1dPlot(resASCA, effectNames = c("Hippurate", "Citrate"))
}
