% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMWLoadingPlot.R
\name{LMWLoadingPlot}
\alias{LMWLoadingPlot}
\title{Loadings plots}
\usage{
LMWLoadingPlot(ResLMWtoPCAEffects, effectNames = NULL, axes = c(1, 2), ...)
}
\arguments{
\item{ResLMWtoPCAEffects}{A list corresponding to the output value of \code{\link{LMWtoPCAEffects}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{...}{Additional arguments to be passed to \code{\link{LinePlot}}.}
}
\value{
A list of loadings plots.
}
\description{
Draws loadings plots for the LMWtoPCAEffects function.
}
\details{
\code{LMWLoadingPlot} is a wrapper of \code{\link{LinePlot}}.
}
\examples{

data('UCH')
ResPCA = SVDforPCA(UCH$outcomes)

PCALoadingPlot(ResSVDforPCA = ResPCA, axes = c(1,2),
title = "PCA loadings plot UCH", xlab = "ppm", ylab = "Values")

}
