% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwModelMatrix.R
\name{lmwModelMatrix}
\alias{lmwModelMatrix}
\title{Creates the model matrix}
\usage{
lmwModelMatrix(lmwDataList)
}
\arguments{
\item{lmwDataList}{A list containing the outcomes, the experimental design and the formula.}
}
\value{
A list with the 5 following named elements :
\describe{
\item{\code{lmwDataList}}{The initial object: a list with outcomes, design and formula.}
\item{\code{modelMatrix}}{A \emph{nxK} model matrix specifically encoded for the ASCA-GLM method.}
\item{\code{modelMatrixByEffect}}{A list of \emph{p} model matrices for each model effect.}
\item{\code{effectsNamesUnique}}{A character vector with the \emph{p} names of the model effects, each repeated once.}
\item{\code{effectsNamesAll}}{A character vector with the \emph{K} names of the model effects ordered and repeated as the column names of the model matrix.}
}
}
\description{
Creates the specific model matrix from the design for the ASCA-GLM method.
}
\details{
In typical ASCA-GLM analysis, the GLM model needs the design to be re-encoded with the commonly used in industrial experimental design \emph{sum coding}. The result is the model matrix.
Suppose the design matrix is \emph{nxk} with n observations and \emph{k} factors. After the transformation, the model matrix
will be \emph{nxK}. For a parameter with \emph{a} levels, the re-encoding is \emph{a-1} columns with 0 and 1 for the \emph{a-1} first levels and -1 for the last one.
\emph{K} is the sum of all the columns for every parameters.
}
\examples{

data('UCH')
resLmwModelMatrix <- lmwModelMatrix(UCH)

head(resLmwModelMatrix$modelMatrix)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
\seealso{
\code{\link{model.matrix}}

More information about the specific encoding is available in the article from (\emph{Thiel et al}, 2017)
}
