% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwPcaEffects.R
\name{lmwPcaEffects}
\alias{lmwPcaEffects}
\title{PCA on the effect matrices}
\usage{
lmwPcaEffects(
  resLmwEffectMatrices,
  method = c("ASCA", "APCA", "ASCA-E"),
  combineEffects = NULL
)
}
\arguments{
\item{resLmwEffectMatrices}{A resLmwEffectMatrices list from \code{\link{lmwEffectMatrices}}.}

\item{method}{The method used to compute the PCA. One of \code{c("ASCA","APCA","ASCA-E")}.}

\item{combineEffects}{If not \code{NULL}, a list of vectors containing the names of the effects to be combined.}
}
\value{
A list of PCA results from \code{\link{pcaBySvd}} for each effect matrix. Those results contain :
\describe{
\item{\code{scores}}{Scores from the PCA for each principal component.}
\item{\code{loadings}}{Loadings from the PCA for each principal component.}
\item{\code{eigval}}{Eigenvalues of each principal component.}
\item{\code{singvar}}{Singular values of each principal component.}
\item{\code{var}}{Explained variances of each principal component.}
\item{\code{cumvar}}{Cumulated explained variances of each principal component.}
\item{\code{original.dataset}}{Original dataset.}
}

There are also others outputs :
\describe{
\item{\code{lmwDataList}}{The initial object: a list of outcomes, design and formula.}
\item{\code{effectsNamesUnique}}{A character vector with the \emph{p} names of the model terms, each repeated once.}
\item{\code{method}}{The dimension reduction method used: \code{c("ASCA","APCA","ASCA-E")}.}
\item{\code{type3SS}}{A vector with the type III SS for each model term.}
\item{\code{variationPercentages}}{A vector with the percentage of variance explained by each model term.}
}
}
\description{
Runs a Principal Component Analysis on the effect matrices and adapts the result according to the method.
}
\details{
The function allows 3 different methods :

\describe{
\item{ASCA}{The PCA is applied directly on the pure effect matrix.}
\item{ASCA-E}{The PCA is applied directly on the pure effect matrix but scores are updated.}
\item{APCA}{The PCA is applied on the augmented effect matrix.}
}
The ASCA-E method adds the residuals to the scores. The APCA method adds the residuals to the effect matrix before the PCA.
}
\examples{
data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resLmwPcaEffects = lmwPcaEffects(resLmwEffectMatrices, method="ASCA-E")
}
