% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMWScorePlot.R
\name{LMWScorePlot}
\alias{LMWScorePlot}
\title{Scores plots}
\usage{
LMWScorePlot(ResLMWtoPCAEffects, effectNames = NULL, axes = c(1, 2), ...)
}
\arguments{
\item{ResLMWtoPCAEffects}{A list corresponding to the output value of \code{\link{LMWtoPCAEffects}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the 2 Principal Components axes to be drawn.}

\item{...}{Additional arguments to be passed to \code{\link{ScatterPlot}}.}
}
\value{
A list of scores plots.
}
\description{
Draws scores plots for the LMWtoPCAEffects function.
}
\details{
\code{LMWScorePlot} is a wrapper of \code{\link{ScatterPlot}}.
}
\examples{

data('UCH')
ResLMWModelMatrix = LMWModelMatrix(UCH)
ResLMWEffectMatrices = LMWEffectMatrices(ResLMWModelMatrix)
ResASCA = LMWtoPCAEffects(ResLMWEffectMatrices)

LMWScorePlot(ResASCA, effectNames = "Hippurate",
color = "Hippurate", shape = "Hippurate")

LMWScorePlot(ResASCA, effectNames = "Hippurate:Time",
color = "Hippurate", shape = "Time")

}
