% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwScoreScatterPlotM.R
\name{lmwScoreScatterPlotM}
\alias{lmwScoreScatterPlotM}
\title{Plotting a Scores Matrix}
\usage{
lmwScoreScatterPlotM(
  resLmwPcaEffects,
  effectNames = NULL,
  PCdim = NULL,
  modelAbbrev = FALSE,
  ...
)
}
\arguments{
\item{resLmwPcaEffects}{A list of p+3 elements depending of the model terms from \code{\link{PCALMEffects}}}

\item{effectNames}{A character vector with the name of the effects to plot}

\item{PCdim}{A numeric vector with the same length than effectNames and indicating the number of component to plot}

\item{modelAbbrev}{A logical whether to abbreviate the interaction terms}

\item{...}{Additional arguments to be passed to \code{\link{plotScatterM}}.}
}
\value{
A matrix of graphs
}
\description{
Plots a matrix of scores graphs
}
\details{
\code{lmwScoreScatterPlotM} is a wrapper of \code{\link{plotScatterM}}.
}
\examples{

 data('UCH')
 ResLMModelMatrix = LMModelMatrix(formula=as.formula(UCH$formula),design=UCH$design)
 ResLMEffectMatrices = LMEffectMatrices(ResLMModelMatrix,outcomes=UCH$outcomes)
 resLmwPcaEffects = PCALMEffects(ResLMEffectMatrices,method="ASCA-E")
 PlotScoresMatrix(resLmwPcaEffects,
                 modelAbbrev=FALSE,
                 design=UCH$design,
                 effectNames = c("Citrate","Hippurate","Hippurate:Citrate"),
                 PCdim=c(2,2,2),
                 varname.colorup = "Citrate",
                 vec.colorup = c("red","blue","green"),
                 varname.pchup="Hippurate",
                 vec.pchup=c(1,2,3),
                 varname.pchdown = "Day",
                 vec.pchdown = c(4,5),
                 varname.colordown="Time",
                 vec.colordown = c("brown","grey"))

}
