% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCALoadingPlot.R
\name{PCALoadingPlot}
\alias{PCALoadingPlot}
\title{Loadings plots}
\usage{
PCALoadingPlot(ResSVDforPCA, axes = c(1, 2), title = "PCA loadings plot", ...)
}
\arguments{
\item{ResSVDforPCA}{A list corresponding to the output value of \code{\link{SVDforPCA}}.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{title}{Plot title.}

\item{...}{Additional arguments to be passed to \code{\link{LinePlot}}.}
}
\value{
A PCA loadings plot.
}
\description{
Draws loadings plots for the SVDforPCA function.
}
\details{
\code{PCALoadingPlot} is a wrapper of \code{\link{LinePlot}}.
}
\examples{

data('UCH')
ResPCA = SVDforPCA(UCH$outcomes)

PCALoadingPlot(ResSVDforPCA = ResPCA, axes = c(1,2),
title = "PCA loadings plot UCH", xlab = "ppm", ylab = "Values")

}
