% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{ScatterPlot}
\usage{
ScatterPlot(
  outcomes,
  design,
  xy,
  color = NULL,
  shape = NULL,
  points_labs = NULL,
  title = "scatterplot",
  xlab = NULL,
  ylab = NULL,
  size = 2,
  size_lab = 3,
  drawShapes = c("none", "ellipse", "polygon", "segment"),
  theme = theme_bw(),
  typeEl = "norm",
  levelEl = 0.9,
  alphaPoly = 0.4
)
}
\arguments{
\item{outcomes}{The nxm matrix with n observations and m response variables.}

\item{design}{The nxk "free encoded" experimental design data frame.}

\item{xy}{x- and y-axis values: A vector of length 2 with either column names of the outcomes matrix to plot (character) or the index positions.}

\item{color}{If not \code{NULL}, a character string giving the column name of \code{design} to be used as color}

\item{shape}{If not \code{NULL}, a character string giving the column name of \code{design} to be used as shape}

\item{points_labs}{If not \code{NULL}, a character vector with point labels.}

\item{title}{Plot title.}

\item{xlab}{If not \code{NULL}, label for the x-axis.}

\item{ylab}{If not \code{NULL}, label for the y-axis.}

\item{size}{The points size.}

\item{size_lab}{The size of points labels.}

\item{drawShapes}{Mulitple shapes can be drawn based on the \code{color}: "none" for non shape (default), "ellipse" (ellipses with ggplot2::stat_ellipse), "polygon" (polygons with ggplot2::geom_polygon) or "segment" (segment from the centroids with ggplot2::geom_segment).}

\item{theme}{ggplot theme, see \code{?ggtheme} for more info.}

\item{typeEl}{The type of ellipse, either \code{norm} (multivariate normal distribution), \code{t} (multivariate t-distribution) or \code{euclid} (draws a circle with the radius equal to level, representing the euclidean distance from the center).}

\item{levelEl}{The confidence level at which to draw an ellipse.}

\item{alphaPoly}{The degree of transparency for polygons.}
}
\value{
A scatterplot in the current device.
}
\description{
Draws a scatter plot.
}
\examples{

data("UCH")

# With color and shape
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
xy = c(453, 369), color = "Hippurate",  shape = "Citrate")

# With color and drawShapes
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
            xy = c(453, 369), color = "Hippurate", 
            drawShapes = "ellipse")

ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
            xy = c(453, 369), color = "Hippurate", 
            drawShapes = "polygon")

ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
            xy = c(453, 369), color = "Hippurate", 
            drawShapes = "segment") 

# customize shapes
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
            xy = c(453, 369), shape = "Hippurate", size = 3) + 
  scale_discrete_identity(aesthetics = 'shape', 
                          guide = 'legend')

ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
            xy = c(453, 369), shape = "Hippurate") + 
  scale_shape_discrete(solid=FALSE)

ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
            xy = c(453, 369), shape = "Hippurate") + 
  scale_shape_manual(values = c(15,16,17))

# with labels
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
xy = c(453, 369), points_labs = rownames(UCH$design))

}
