% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVDforPCA.R
\name{SVDforPCA}
\alias{SVDforPCA}
\title{Singular value decomposition for PCA analysis}
\usage{
SVDforPCA(outcomes, ncomp = min(dim(outcomes)))
}
\arguments{
\item{outcomes}{The nxm matrix with n observations and m response variables.}

\item{ncomp}{Number of Principal Components.}
}
\value{
A list with the following elements:
\describe{
\item{\code{scores}}{Scores}
\item{\code{loadings}}{Loadings}
\item{\code{eigval}}{Eigenvalues}
\item{\code{pcd}}{Singular values}
\item{\code{pcu}}{Normalized scores}
\item{\code{var}}{Explained variance}
\item{\code{cumvar}}{Cumulated explained variance}
\item{\code{original.dataset}}{Original dataset}

}
}
\description{
PCA by singular value decomposition, the preprocessing involves the mean-centering of X.
}
\examples{
data('UCH')
PCA.res = SVDforPCA(UCH$outcomes)
}
