% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmpScreePlot.R
\name{lmpScreePlot}
\alias{lmpScreePlot}
\title{Scree Plot}
\usage{
lmpScreePlot(
  resLLmpContributions,
  effectNames = NULL,
  nPC = 5,
  theme = theme_bw()
)
}
\arguments{
\item{resLLmpContributions}{A resLLmpContributions list from the function \code{\link{lmpContributions}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{nPC}{An integer with the number of components to plot.}

\item{theme}{\code{ggplot} theme}
}
\value{
A scree plot (ggplot).
}
\description{
Plots barplots of the percentage of variance associated with each principal component and for each matrix effect on the basis of \code{\link{lmpContributions}} outputs.
}
\examples{

 data('UCH')
 resLLmpModelMatrix = lmpModelMatrix(UCH)
 resLLmpEffectMatrices = lmpEffectMatrices(resLLmpModelMatrix)
 resASCAE = lmpPcaEffects(resLLmpEffectMatrices, method="ASCA-E")
 resLLmpContributions = lmpContributions(resASCAE)
 lmpScreePlot(resLLmpContributions, effectNames ="Hippurate:Citrate", nPC=4)

}
