% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwBootstrapTests.R
\name{lmwBootstrapTests}
\alias{lmwBootstrapTests}
\title{Performs a test on the effects from the model}
\usage{
lmwBootstrapTests(resLmwEffectMatrices, nboot = 100, nCores = 2)
}
\arguments{
\item{resLmwEffectMatrices}{A list of 11 from \code{\link{lmwEffectMatrices}}}

\item{nboot}{A integer with the number of iterations to perform.}

\item{nCores}{The number of cores to use for parallel execution.}
}
\value{
A list with the following elements:
\describe{
\item{\code{Fobs}}{A vector with the F statistics observed with the data}
\item{\code{Fboot}}{A matrix with the F statistics observed with the bootstrap}
\item{\code{Pvalues}}{A vector with the pvalue for every effect}
\item{\code{ResultsTable}}{A results matrix with the pvalue and the variation percentage for every effect}
}
}
\description{
Computes a partial model without the tested effects then estimates the residuals. Next, computes new outcomes from the predicted values of the partial model and sampled residuals. Finally, computes the Sum of Squares and the test statistics.
}
\details{
The function works as follow:

To be written again
}
\examples{
 data('UCH')
 resLmwModelMatrix <- lmwModelMatrix(UCH)
 resLmwEffectMatrices <- lmwEffectMatrices(resLmwModelMatrix = resLmwModelMatrix)

 res <- lmwBootstrapTests(resLmwEffectMatrices = resLmwEffectMatrices, nboot=10, nCores=2)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
