% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwContributions.R
\name{lmwContributions}
\alias{lmwContributions}
\title{Summary of the contributions from each effect}
\usage{
lmwContributions(resLmwPcaEffects, nPC = 5)
}
\arguments{
\item{resLmwPcaEffects}{A resLmwPcaEffects list from \code{\link{lmwPcaEffects}}}

\item{nPC}{The number of Principal Components to print in the tables}
}
\value{
A list of 3 :
\describe{
\item{\code{EffectTable}}{A matrix with the variance percentage of each PC for each effect}
\item{\code{ContribTable}}{A matrix with the contribution of each PC to the total variance}
\item{\code{Barplot}}{A barplot with the PCs which have the biggest contributions to the total variance}
}
}
\description{
Creates a summary of the contribution of each effect and each of their Principal Component (PC) on the total variance. Additionally plots a graph with the ordered contributions.
}
\examples{
data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resLmwPcaEffects = lmwPcaEffects(resLmwEffectMatrices, method="ASCA-E")

lmwContributions(resLmwPcaEffects)

}
