% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwContributions.R
\name{lmwContributions}
\alias{lmwContributions}
\title{Summary of the contributions of each effect}
\usage{
lmwContributions(resLmwPcaEffects, nPC = 5)
}
\arguments{
\item{resLmwPcaEffects}{A list corresponding to the output value of \code{\link{lmwPcaEffects}}.}

\item{nPC}{The number of Principal Components to display.}
}
\value{
A list of:
\describe{
\item{\code{totalContribTable}}{Table of the percentage of contribution of each effect to the total variance.}
\item{\code{effectTable}}{Table of the variance percentage explained by each Principal Component in each model effect decomposition.}
\item{\code{contribTable}}{Table of the variance percentage explained by each Principal Component of each effect reported to the percentage contribution of the given effect to the total variance.}
\item{\code{combinedEffectTable}}{Equivalent of the \emph{EffectTable} for combined effects.}
\item{\code{plotTotal}}{Plot of the ordered contributions of \emph{TotalContribTable}.}
\item{\code{plotContrib}}{Plot of the ordered contributions of \emph{ContribTable}.}
}
}
\description{
Reports the contribution of each effect to the total variance, but also the contribution of each principal component to the total variance per effect. Moreover, these contributions are summarized in a barplot.
}
\examples{
data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resLmwPcaEffects = lmwPcaEffects(resLmwEffectMatrices, method="ASCA-E")

lmwContributions(resLmwPcaEffects)

}
