% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMWEffectMatrices.R
\name{LMWEffectMatrices}
\alias{LMWEffectMatrices}
\title{Computing the Effect Matrices}
\usage{
LMWEffectMatrices(
  ResLMWModelMatrix,
  SS = TRUE,
  newSSmethod = TRUE,
  contrastList = NA
)
}
\arguments{
\item{ResLMWModelMatrix}{A list of 5 elements from \code{\link{LMWModelMatrix}}.}

\item{SS}{Logical. If \code{FALSE}, won't compute the effect percentage variations.}

\item{newSSmethod}{A logical whether to use the new optimized method to compute SS.}

\item{contrastList}{A list of contrast for each parameter. The function creates automatically the list by default.}
}
\value{
A list with the following elements:
\describe{
\item{\code{LMWiRe_data_list}}{A list containing the outcomes, the experimental design and the formula.}
\item{\code{ModelMatrix}}{A \emph{nxK} model matrix specifically encoded for the ASCA-GLM method.}
\item{\code{modelMatrixByEffect}}{A list of \emph{p} model matrices by models terms.}
\item{\code{covariateEffectsNames}}{A character vector with \emph{K} names of the coefficients.}
\item{\code{covariateEffectsNamesUnique}}{A character vector with the \emph{p} unique names of the model terms.}
\item{\code{effectMatrices}}{A list of \emph{p} effect matrices for each model terms.}
\item{\code{predictedvalues}}{A \emph{nxm} matrix with the predicted values.}
\item{\code{residuals}}{A \emph{nxm} matrix with the residuals.}
\item{\code{parameters}}{A \emph{pxm} matrix with the coefficients of every parameters by response variables.}
\item{\code{SS}}{A vector with the type III residuals for each model terms.}
\item{\code{variationPercentages}}{A \emph{p} named vector with the variation percentages of each model terms.}
}
}
\description{
Runs a GLM model and decomposes the outcomes into effect matrices for each model terms
}
\examples{
 data('UCH')
 ResLMWModelMatrix <- LMWModelMatrix(UCH)
 LMWEffectMatrices(ResLMWModelMatrix)

}
