% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwEffectPlot.R
\name{lmwEffectPlot}
\alias{lmwEffectPlot}
\title{Effect plot}
\usage{
lmwEffectPlot(
  resASCA,
  effectName,
  axes = 1,
  x,
  z = NULL,
  w = NULL,
  hline = 0,
  ...
)
}
\arguments{
\item{resASCA}{A list corresponding to the ASCA output value of \code{\link{lmwPcaEffects}}.}

\item{effectName}{Name of the effect matrix used for the scores.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{x}{A character string giving the \code{design} factor whose levels will form the x axis.}

\item{z}{A character string giving the \code{design} factor whose levels will form the traces.}

\item{w}{A character string giving the \code{design} factor whose levels will be used for the facet.}

\item{hline}{If not \code{NULL}, draws (a) horizontal line(s).}

\item{...}{Additional arguments to be passed to \code{\link{plotMeans}}.}
}
\value{
An effect plot (ggplot).
}
\description{
Plots the ASCA scores by effect levels for a given model effect (for one by principal component at a time). This graph is particularly appealing to interpret interactions or combined effects.
}
\details{
\code{lmwEffectPlot} is a wrapper of \code{\link{plotMeans}}.
}
\examples{

  resLmwModelMatrix = lmwModelMatrix(UCH)
  resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
  resASCA = lmwPcaEffects(resLmwEffectMatrices = resLmwEffectMatrices,
method="ASCA", combineEffects = list(c("Hippurate", "Time",  "Hippurate:Time")))
  lmwEffectPlot(resASCA, effectName = "Hippurate", x = "Hippurate")

}
