% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwLoading2dPlot.R
\name{lmwLoading2dPlot}
\alias{lmwLoading2dPlot}
\title{Loading plots}
\usage{
lmwLoading2dPlot(
  resLmwPcaEffects,
  effectNames = NULL,
  axes = c(1, 2),
  metadata = NULL,
  ...
)
}
\arguments{
\item{resLmwPcaEffects}{A list corresponding to the output value of \code{\link{lmwPcaEffects}}.}

\item{effectNames}{Names of the effects to be plotted. If \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the 2 Principal Components axes to be drawn.}

\item{metadata}{A nxk "free encoded" data.frame corresponding to \code{design} in \code{\link{plotScatter}}.}

\item{...}{Additional arguments to be passed to \code{\link{plotScatter}}.}
}
\value{
A list of loading plots (ggplot).
}
\description{
Draws the Loading plots of each effect matrix provided in \code{\link{lmwPcaEffects}} outputs. As a wrapper of the \code{\link{plotScatter}} function, it allows the visualization of effect loading matrices for two components at a time with all options available in \code{\link{plotScatter}}.
}
\details{
\code{lmwLoading2dPlot} is a wrapper of \code{\link{plotScatter}}.
}
\examples{

data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resASCA = lmwPcaEffects(resLmwEffectMatrices)

labels = substr(colnames(UCH$outcomes),1,4)
ids <- order(resASCA$Hippurate$loadings[,1], decreasing = TRUE)[1:10]
labels[-c(ids)] <- ""
groups <- rep(c(1,2), length.out = ncol(UCH$outcomes))
metadata <- data.frame(groups)
lmwLoading2dPlot(resASCA,effectNames = "Hippurate",
metadata=metadata, color = "groups", shape = "groups",
points_labs = labels)
}
