% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMWLoadingPlot.R
\name{lmwLoadingPlot}
\alias{lmwLoadingPlot}
\title{Loadings plots}
\usage{
lmwLoadingPlot(resLmwPcaEffects, effectNames = NULL, axes = c(1, 2), ...)
}
\arguments{
\item{resLmwPcaEffects}{A list corresponding to the output value of \code{\link{lmwPcaEffects}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{...}{Additional arguments to be passed to \code{\link{plotLine}}.}
}
\value{
A list of loadings plots.
}
\description{
Draws loadings plots for the lmwPcaEffects function.
}
\details{
\code{lmwLoadingPlot} is a wrapper of \code{\link{plotLine}}.
}
\examples{

data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resASCA = lmwPcaEffects(resLmwEffectMatrices)

lmwLoadingPlot(resASCA, effectNames = c("Hippurate", "Citrate"))
}
