% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMWModelMatrix.R
\name{LMWModelMatrix}
\alias{LMWModelMatrix}
\title{Creating the model matrix}
\usage{
LMWModelMatrix(LMWiRe_data_list)
}
\arguments{
\item{LMWiRe_data_list}{A list containing the outcomes, the experimental design and the formula.}
}
\value{
A list with the 5 following named elements :
\describe{
\item{\code{LMWiRe_data_list}}{A \emph{nxk} data frame with the "free encoded" experimental design.}
\item{\code{ModelMatrix}}{A \emph{nxK} model matrix specifically encoded for the ASCA-GLM method.}
\item{\code{modelMatrixByEffect}}{A list of \emph{p} model matrices by models terms.}
\item{\code{covariateEffectsNames}}{A character vector with \emph{K} names of the coefficients.}
\item{\code{covariateEffectsNamesUnique}}{A character vector with the \emph{p} unique names of the model terms.}
}
}
\description{
Create the specific model matrix from the design for the ASCA-GLM method.
}
\details{
In typical ASCA-GLM analysis, the GLM model needs the design to be re-encoded with the commonly used in industrial experimental design \emph{sum coding}. The result is the model matrix.
Suppose the design matrix is \emph{nxk} with n observations and \emph{k} factors. After the transformation the model matrix
will be \emph{nxK}. For a parameter with a levels the re-encoding is \emph{a-1} colums with 0 and 1 for \emph{a-1} first levels and -1 for the last.
\emph{K} is the sum of all the columns for every parameters.
}
\examples{

data('UCH')

ResLMWModelMatrix <- LMWModelMatrix(UCH)

head(ResLMWModelMatrix$ModelMatrix)

}
\references{
Thiel M.,Feraud B. and Govaerts B. (2017) \emph{ASCA+ and APCA+: Extensions of ASCA and APCA
in the analysis of unbalanced multifactorial designs}, Journal of Chemometrics
}
\seealso{
\code{\link{model.matrix}}

More information about the specific encoding is available in the article from (\emph{Thiel et al}, 2017)
}
