% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwPcaEffects.R
\name{lmwPcaEffects}
\alias{lmwPcaEffects}
\title{PCA on the effect matrices}
\usage{
lmwPcaEffects(
  resLmwEffectMatrices,
  method = c("ASCA", "APCA", "ASCA-E"),
  combineEffects = NULL
)
}
\arguments{
\item{resLmwEffectMatrices}{A resLmwEffectMatrices list from \code{\link{lmwEffectMatrices}}}

\item{method}{The method used to compute the PCA. One of \code{c("ASCA","APCA","ASCA-E")}}

\item{combineEffects}{If not \code{NULL}, a list of vectors containing the names of the effects to be combined.}
}
\value{
A list of PCA results from \code{\link{pcaBySvd}} for each effect matrix. Those results contain :
\describe{
\item{\code{scores}}{Scores from the PCA for each of the n components.}
\item{\code{loadings}}{Loadings from the PCA for each of the n components.}
\item{\code{eigval}}{Eigenvalues of each of the n components.}
\item{\code{pcu}}{\emph{nxn} matrix of normalized scores.}
\item{\code{pcd}}{Singular values of each of the n components.}
\item{\code{var}}{Explained variance of each of the n components.}
\item{\code{cumvar}}{Cumulated explained variance of each of the n components.}
\item{\code{original.dataset}}{Original dataset}
}
There are also others outputs :
\describe{
\item{\code{lmwDataList}}{A list containing the outcomes, the experimental design and the formula.}
\item{\code{covariateEffectsNamesUnique}}{A character vector with the \emph{p} unique names of the model terms.}
\item{\code{method}}{The method used to compute the PCA. One of \code{c("ASCA","APCA","ASCA-E").}
\item{\code{SS}}{A list of matrices with the type III residuals for each model terms.}
\item{\code{variationPercentages}}{A \emph{p} named vector with the variation percentages of each model terms.}
}
}
\description{
Runs a Principal Component Analysis on the effect matrices and adapts the result according to the method.
}
\details{
The function allows 3 different methods :

\describe{
\item{ASCA}{The PCA is applied directly on the pure effect matrix.}
\item{ASCA-E}{The PCA is applied directly on the pure effect matrix but scores are updated.}
\item{APCA}{The PCA is applied on the augmented effect matrix.}
}
The ASCA-E method adds the residuals to the scores. The APCA method adds the residuals to the effect matrix before the PCA.
}
\examples{
data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resLmwPcaEffects = lmwPcaEffects(resLmwEffectMatrices, method="ASCA-E")

PlotScoresXY(resLmwPcaEffects, UCH$design, EffectVector=c("Hippurate"),
             varname.color=c("Citrate"), varname.pch=c("Time"))
}
