% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwScorePlot.R
\name{lmwScorePlot}
\alias{lmwScorePlot}
\title{Scores plots}
\usage{
lmwScorePlot(resLmwPcaEffects, effectNames = NULL, axes = c(1, 2), ...)
}
\arguments{
\item{resLmwPcaEffects}{A list corresponding to the output value of \code{\link{lmwPcaEffects}}.}

\item{effectNames}{Names of the effects to be plotted. If \code{NULL}, all the effects are plotted.}

\item{axes}{A numerical vector with the 2 Principal Components axes to be drawn.}

\item{...}{Additional arguments to be passed to \code{\link{plotScatter}}.}
}
\value{
A list of scores plots (ggplot).
}
\description{
Draws scores plots for the \code{\link{lmwPcaEffects}} function.
}
\details{
\code{lmwScorePlot} is a wrapper of \code{\link{plotScatter}}.
}
\examples{

data('UCH')
resLmwModelMatrix = lmwModelMatrix(UCH)
resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
resASCA = lmwPcaEffects(resLmwEffectMatrices)

lmwScorePlot(resASCA, effectNames = "Hippurate",
color = "Hippurate", shape = "Hippurate")

lmwScorePlot(resASCA, effectNames = "Hippurate:Time",
color = "Hippurate", shape = "Time")

}
