% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwScreePlot.R
\name{lmwScreePlot}
\alias{lmwScreePlot}
\title{Scree Plot}
\usage{
lmwScreePlot(
  resLmwPcaEffects,
  effectName,
  title = paste0("Scree plot of the ", effectName, " effect"),
  nPC = 5
)
}
\arguments{
\item{resLmwPcaEffects}{A resLmwPcaEffects list from the function \code{\link{lmwPcaEffects}}.}

\item{effectName}{The name of the effect to be plotted.}

\item{title}{Plot title.}

\item{nPC}{An integer with the number of components to plot.}
}
\value{
A scree plot
}
\description{
Draws scree plots for the lmwPcaEffects function.
}
\examples{

 data('UCH')
 resLmwModelMatrix = lmwModelMatrix(UCH)
 resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
 resLmwPcaEffects = lmwPcaEffects(resLmwEffectMatrices,method="ASCA-E")
 lmwScreePlot(resLmwPcaEffects,"Hippurate:Citrate",nPC=4)

}
