% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmwScreePlot.R
\name{lmwScreePlot}
\alias{lmwScreePlot}
\title{Scree Plot}
\usage{
lmwScreePlot(
  resLmwContributions,
  effectNames = NULL,
  nPC = 5,
  theme = theme_bw()
)
}
\arguments{
\item{resLmwContributions}{A resLmwContributions list from the function \code{\link{lmwContributions}}.}

\item{effectNames}{Names of the effects to be plotted. if \code{NULL}, all the effects are plotted.}

\item{nPC}{An integer with the number of components to plot.}
}
\value{
A scree plot (ggplot).
}
\description{
Draws scree plots for the \code{\link{lmwContributions}} function.
}
\examples{

 data('UCH')
 resLmwModelMatrix = lmwModelMatrix(UCH)
 resLmwEffectMatrices = lmwEffectMatrices(resLmwModelMatrix)
 resASCAE = lmwPcaEffects(resLmwEffectMatrices, method="ASCA-E")
 resLmwContributions = lmwContributions(resASCAE)
 lmwScreePlot(resLmwContributions, effectNames ="Hippurate:Citrate", nPC=4)

}
