% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaBySvd.R
\name{pcaBySvd}
\alias{pcaBySvd}
\title{Singular value decomposition for PCA analysis}
\usage{
pcaBySvd(Y, nPC = min(dim(Y)))
}
\arguments{
\item{Y}{The nxm matrix with n observations and m response variables.}

\item{nPC}{Number of Principal Components.}
}
\value{
A list with the following elements:
\describe{
\item{\code{scores}}{Scores}
\item{\code{loadings}}{Loadings}
\item{\code{eigval}}{Eigenvalues}
\item{\code{singvar}}{Singular values}
\item{\code{var}}{Explained variances}
\item{\code{cumvar}}{Cumulated explained variances}
\item{\code{original.dataset}}{Original dataset}

}
}
\description{
PCA by singular value decomposition, the pre-processing involves the mean-centering of X.
}
\examples{

data('UCH')
PCA.res = pcaBySvd(UCH$outcomes)
}
