% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaLoading2dPlot.R
\name{pcaLoading2dPlot}
\alias{pcaLoading2dPlot}
\title{Loading plots on a 2D scatter plot}
\usage{
pcaLoading2dPlot(
  resPcaBySvd,
  axes = c(1, 2),
  title = "PCA loading plot",
  points_labs_rn = FALSE,
  metadata = NULL,
  ...
)
}
\arguments{
\item{resPcaBySvd}{A list corresponding to the output value of \code{\link{pcaBySvd}}.}

\item{axes}{A numerical vector with the 2 Principal Components axes to be drawn.}

\item{title}{Plot title.}

\item{points_labs_rn}{Boolean indicating if the rownames of the loadings matrix should be plotted.}

\item{metadata}{A nxk "free encoded" data.frame corresponding to \code{design} in \code{\link{plotScatter}}.}

\item{...}{Additional arguments to be passed to \code{\link{plotScatter}}.}
}
\value{
A \code{ggplot} object with the PCA loading plot.
}
\description{
Produces loading plots from \code{\link{pcaBySvd}} with the same graphical options as \code{\link{plotScatter}}.
}
\details{
\code{pcaLoading2dPlot} is a wrapper of \code{\link{plotScatter}}.
}
\examples{

data('UCH')
ResPCA = pcaBySvd(UCH$outcomes)

# adding (arbitrary) color and shape to points
groups <- rep(c(1,2), length.out = ncol(UCH$outcomes))
metadata <- data.frame(groups)

pcaLoading2dPlot(resPcaBySvd = ResPCA, axes = c(1,2),
title = "PCA loading plot UCH", metadata = metadata,
color = "groups", shape = "groups")

}
