% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaLoadingPlot.R
\name{pcaLoadingPlot}
\alias{pcaLoadingPlot}
\title{Loading plots}
\usage{
pcaLoadingPlot(resPcaBySvd, axes = c(1, 2), title = "PCA loading plot", ...)
}
\arguments{
\item{resPcaBySvd}{A list corresponding to the output value of \code{\link{pcaBySvd}}.}

\item{axes}{A numerical vector with the Principal Components axes to be drawn.}

\item{title}{Plot title.}

\item{...}{Additional arguments to be passed to \code{\link{plotLine}}.}
}
\value{
A PCA loading plot (ggplot).
}
\description{
Plots the loadings from \code{\link{pcaBySvd}} with different possible line types.
}
\details{
\code{pcaLoadingPlot} is a wrapper of \code{\link{plotLine}}.
}
\examples{

data('UCH')
ResPCA = pcaBySvd(UCH$outcomes)

pcaLoadingPlot(resPcaBySvd = ResPCA, axes = c(1,2),
title = "PCA loading plot UCH", xlab = "ppm", ylab = "Values")
}
