% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAScorePlot.R
\name{PCAScorePlot}
\alias{PCAScorePlot}
\title{Scores plots}
\usage{
PCAScorePlot(
  ResSVDforPCA,
  axes = c(1, 2),
  title = "PCA scores plot",
  points_labs_rn = FALSE,
  ...
)
}
\arguments{
\item{ResSVDforPCA}{A list corresponding to the output value of \code{\link{SVDforPCA}}.}

\item{axes}{A numerical vector with the 2 Principal Components axes to be drawn.}

\item{title}{Plot title.}

\item{points_labs_rn}{Boolean indicating if the rownames of the scores matrix should be plotted.}

\item{...}{Additional arguments to be passed to \code{\link{ScatterPlot}}.}
}
\value{
A PCA scores plot.
}
\description{
Draws scores plots for the SVDforPCA function.
}
\details{
\code{PCAScorePlot} is a wrapper of \code{\link{ScatterPlot}}.
}
\examples{

data('UCH')
ResPCA = SVDforPCA(UCH$outcomes)

PCAScorePlot(ResSVDforPCA = ResPCA, axes = c(1,2), 
title = "PCA scores plot UCH", design = UCH$design,
color = "Hippurate", shape = "Citrate")

}
