% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaScorePlot.R
\name{pcaScorePlot}
\alias{pcaScorePlot}
\title{Scores plots}
\usage{
pcaScorePlot(
  resPcaBySvd,
  axes = c(1, 2),
  title = "PCA scores plot",
  points_labs_rn = FALSE,
  ...
)
}
\arguments{
\item{resPcaBySvd}{A list corresponding to the output value of \code{\link{pcaBySvd}}.}

\item{axes}{A numerical vector with the 2 Principal Components axes to be drawn.}

\item{title}{Plot title.}

\item{points_labs_rn}{Boolean indicating if the rownames of the scores matrix should be plotted.}

\item{...}{Additional arguments to be passed to \code{\link{plotScatter}}.}
}
\value{
A PCA scores plot.
}
\description{
Draws scores plots for the pcaBySvd function.
}
\details{
\code{pcaScorePlot} is a wrapper of \code{\link{plotScatter}}.
}
\examples{

data('UCH')
ResPCA = pcaBySvd(UCH$outcomes)

pcaScorePlot(resPcaBySvd = ResPCA, axes = c(1,2),
title = "PCA scores plot UCH", design = UCH$design,
color = "Hippurate", shape = "Citrate")

}
