% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDesign.R
\name{plotDesign}
\alias{plotDesign}
\title{Plot of the design matrix}
\usage{
plotDesign(
  design,
  x,
  y,
  rows = NULL,
  cols = NULL,
  title = "Plot of the design",
  theme = theme_bw()
)
}
\arguments{
\item{design}{A nxk "free encoded" experimental design data frame.}

\item{x}{A character string giving the column name of \code{design} to be used for the x axis.}

\item{y}{A character string giving the column name of \code{design} to be used for the y axis.}

\item{rows}{If not \code{NULL}, a character vector with one or several column name(s) of \code{design} to be used for faceting along the rows.}

\item{cols}{If not \code{NULL}, a character vector with one or several column name(s) of \code{design} to be used for faceting along the columns.}

\item{title}{Plot title.}

\item{theme}{ggplot theme, see \code{?ggtheme} for more info.}
}
\value{
A plot of the design (ggplot).
}
\description{
Plots the experimental design of the data.
}
\examples{

### mtcars
data(mtcars)
df <-  mtcars \%>\%
  select(cyl, vs, am, gear, carb) \%>\%
  as.data.frame() \%>\%
  mutate(across(everything(), as.factor))

# 2 factors
plotDesign(design = df, x = "cyl", y = "vs", rows = NULL,
          cols = NULL)
# 3 factors
plotDesign(design = df, x = "cyl", y = "vs", rows = c("am"),
          cols = NULL)
# 4 factors
plotDesign(design = df, x = "cyl", y = "vs", rows = c("am"),
          cols = c("gear"))
# 5 factors
plotDesign(design = df, x = "cyl", y = "vs", rows = c("am","carb"),
          cols = c("gear"))

plotDesign(design = df, x = "cyl", y = "vs", rows = c("am","carb"),
          cols = c("vs"))

### UCH
data("UCH")
plotDesign(design = UCH$design, x = "Hippurate", y = "Citrate", rows = "Day")

}
