% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLine.R
\name{plotLine}
\alias{plotLine}
\title{Line plot}
\usage{
plotLine(
  X,
  title = "Line plot",
  rows = 1,
  type = c("l", "p", "s"),
  xlab = NULL,
  ylab = NULL,
  xaxis_type = c("numeric", "character"),
  stacked = FALSE,
  ncol = 1,
  nrow = NULL,
  facet_label = NULL,
  hline = 0,
  size = 0.5,
  color = NULL,
  shape = 1,
  theme = theme_bw()
)
}
\arguments{
\item{X}{A numerical matrix containing the rows to be drawn.}

\item{title}{Plot title.}

\item{rows}{A vector with either the row name(s) of the X matrix to plot (character) or the row index position(s).}

\item{type}{Type of graph to be drawn: "p" for point, "l" for line or "s" for segment.}

\item{xlab}{If not \code{NULL}, label for the x-axis.}

\item{ylab}{If not \code{NULL}, label for the y-axis.}

\item{xaxis_type}{The data type of the x axis: either "numeric" or "character".}

\item{stacked}{Logical. If \code{TRUE}, will draw stacked plots, otherwise will draw separate plots.}

\item{ncol}{If stacked is \code{FALSE}, the number of columns to represent the separate plots.}

\item{nrow}{If stacked is \code{FALSE}, the number of rows to represent the separate plots.}

\item{facet_label}{If stacked is \code{FALSE}, the labels of the separate plots.}

\item{hline}{If not \code{NULL}, draws (a) horizontal line(s).}

\item{size}{Argument of length 1 giving the points size (if \code{type} == "p") or the line size (if \code{type} == "l" or "s").}

\item{color}{If not \code{NULL}, argument of length 1 with possible values: "rows", a color name (character) or a numeric value representing a color.}

\item{shape}{The points shape if \code{type} == "p" (argument of length 1).}

\item{theme}{ggplot theme, see \code{?ggtheme} for more info.}
}
\value{
A line plot.
}
\description{
Draws Line plot(s).
}
\examples{

plotLine(X = UCH$outcomes)

# separate plots
plotLine(X = UCH$outcomes, rows = c(1:4), hline = NULL)
plotLine(X = UCH$outcomes, rows = c(1:4), color = 2)
plotLine(X = UCH$outcomes, rows = c(1:8), ncol=2)
plotLine(X = UCH$outcomes, type = "p",
         rows = c(1:8), ncol=2)

# stacked plots
plotLine(X = UCH$outcomes, rows = c(1:4),
         stacked = TRUE, color = "rows") +
         scale_color_brewer(palette="Set1")

}
