% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{ScatterPlot}
\usage{
ScatterPlot(
  outcomes,
  design,
  xy,
  color = NULL,
  shape = NULL,
  points_labs = NULL,
  title = "scatterplot",
  xlab = NULL,
  ylab = NULL,
  size = 2,
  size_lab = 3,
  drawEllipses = FALSE,
  typeEl = c("norm", "t", "euclid"),
  levelEl = 0.9
)
}
\arguments{
\item{outcomes}{The nxm matrix with n observations and m response variables.}

\item{design}{The nxk "free encoded" experimental design data frame.}

\item{xy}{x- and y-axis values: A vector of length 2 with either column names of the outcomes matrix to plot (character) or the index positions.}

\item{color}{If not \code{NULL}, a character string giving the column name of \code{design} to be used as color}

\item{shape}{If not \code{NULL}, a character string giving the column name of \code{design} to be used as shape}

\item{points_labs}{If not \code{NULL}, a character vector with point labels.}

\item{title}{Plot title.}

\item{xlab}{If not \code{NULL}, label for the x-axis.}

\item{ylab}{If not \code{NULL}, label for the y-axis.}

\item{size}{The points size.}

\item{size_lab}{The size of points labels.}

\item{drawEllipses}{If \code{TRUE}, will draw ellipse(s).}

\item{typeEl}{The type of ellipse, either \code{norm} (multivariate normal distribution), \code{t} (multivariate t-distribution) or \code{euclid} (draws a circle with the radius equal to level, representing the euclidean distance from the center).}

\item{levelEl}{The confidence level at which to draw an ellipse.}
}
\value{
A scatterplot in the current device.
}
\description{
Draws a scatter plot.
}
\examples{

data("UCH")

# With color and shape
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
xy = c(453, 369), color = "Hippurate",  shape = "Citrate")

# With color and ellipses
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
xy = c(453, 369), color = "Hippurate", 
drawEllipses = TRUE)

# with labels
ScatterPlot(outcomes = UCH$outcomes, design = UCH$design, 
xy = c(453, 369), points_labs = rownames(UCH$design))

}
